test_api/requirements.txt
# 核心依赖
pytest>=7.4.0
pytest-html>=3.2.0
pytest-xdist>=3.3.0
pytest-rerunfailures>=12.0
pytest-ordering>=0.6
allure-pytest>=2.13.0
requests>=2.31.0
PyYAML>=6.0.1

# 数据驱动
openpyxl>=3.1.2

# 日志和报告
Pillow>=10.0.0

# 代码质量
flake8>=6.1.0
black>=23.7.0

# 类型检查
mypy>=1.5.0

# 环境管理（必需）
python-dotenv>=1.0.0
